feat: removes viper for config loading and uses JSON standard library
Due to inconsistencies between the way viper and the JSON standard library unmarshals
number values between float/int, removing in favor of sticking to the standard lib to have
consistent attribute sets.

+++ attributes/bool.go
@@ -51,7 +51,7 @@ func (a boolAttribute) AsFloat() (float64, error) {
 
 // AsInt returns the value as an int value errors and if that is not
 // the underlying type.
-func (a boolAttribute) AsInt() (int, error) {
+func (a boolAttribute) AsInt() (int64, error) {
 	return 0, ErrWrongKind
 }
